version 1.0.3 - XMLHttpRequest.prototype.send test
OK, this is promising.  Here is the idea.  We override
XMLHttpRequest.prototype.send.  When the client is sending
html content, we purify it before sending. For example in an
email that is being sent, we see the following:

{
  "Body": {
    "SendMsgRequest": {
      "m": {
        "mp": [
           {
             "content": {
               "_content": "....."
             },
             "ct": "CONTENT/TYPE"
           },
           ...

For all of the _content's whose corresponding ct's are
"text/html", we could purify them before sending.